refactor: TransactionDispatcher를 추상화시키고, 로직을 따르도록 강제화 한다
